function array

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -4,14 +4,14 @@
 /**
  * array_iterator - function poiniter
  * @array: char
- * @size_t: size
+ * @size: size
  * @action: char function
  * Return: pointer
  */
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	long unsigned int i = 0;
+	unsigned int i = 0;
 
 	while (i < size)
 	{
